fix global
