+++ src/context/GlobalContext.jsx
@@ -5,7 +5,7 @@ import { createContext, useState } from 'react';
 export const GlobalContext = createContext();
 
 export const GlobalProvider = ({ children }) => {
-    const [user, setUser] = useState(null);
+    const [user, setUser] = useState("ole");
 
     return (
         <GlobalContext.Provider value={{ user, setUser }}>
