+++ src/main.jsx
@@ -1,14 +1,3 @@
 import { createRoot } from 'react-dom/client'
-import './index.css'
-import App from './App.jsx'
-import { BrowserRouter } from "react-router";
-import { GlobalProvider } from "../context/GlobalContext.jsx";
-
-createRoot(document.getElementById('root')).render(
-  <BrowserRouter>
-    <GlobalProvider>
-      <App />
-    </GlobalProvider>
-  </BrowserRouter>
-
-)
+import Routes from './routes.jsx';
+createRoot(document.getElementById('root')).render(<Routes />)
