+++ src/page/About.jsx
@@ -1,11 +1,9 @@
-import { Link, NavLink } from "react-router";
-import Navaiagtion from "../Naviagtion";
 
-export default function About(){
-    return(
+export default function About() {
+    return (
         <section>
             <h1>About</h1>
         </section>
     )
-} 
+}
