+++ src/page/Home.jsx
@@ -1,10 +1,17 @@
-//import { FaBeer } from 'react-icons/fa';
+import { useContext, useEffect } from "react"
+import { GlobalContext } from "../context/GlobalContext"
+
+
 export default function Home() {
+    const { user, setUser } = useContext(GlobalContext)
+
     return (
+
         <section>
-            <h1>Home  </h1>
+            <h1>Velkommen  {user && user}</h1>
+            <button onClick={() => {
+                setUser("Steen")
+            }}>Tilføj bruger navn</button>
         </section>
     )
 }
-
-{/* <FaBeer color="#999" /> */ }
